v0.1
The script initialises Chrome WebDriver and opens the YouTube homepage.

Wait for the YouTube search box to load, then type the title "Aalto Talk with Linus Torvalds" verbatim, with a slight delay between each character to mimic human typing. After a short random wait, the search button is clicked.

The script waits for the search results to load, then scrolls down the page.

It finds the video called "Aalto Talk with Linus Torvalds [Full-length]" and clicks on it. After waiting for the video (and any adverts) to load, it tries to skip the adverts if there are any.

The script will wait a certain amount of time before liking the video. It will then scroll slightly, wait for the comment box to load, type a comment ("Nice GG!"), and submit the comment after a short delay.

Finally, the script closes the browser.